pdo & pdoext classes merged
