+++ src/App/bootstrap.php
@@ -14,8 +14,9 @@ use App\RequestHandler\Logout;
 use Clear\Config\Factory as ConfigFactory;
 use Clear\Config\ConfigInterface;
 use Clear\Container\Container;
-use Clear\Database\Pdo as PDO;
+use Clear\Database\PdoExt as PDO;
 use Clear\Database\Event\AfterConnect;
+use Clear\Database\PdoInterface;
 use Clear\Events\Dispatcher;
 use Clear\Events\Provider;
 use Clear\Http\Router;
@@ -75,7 +76,7 @@ $app->eventDispatcher = function () use ($app) {
 };
 
 // Database connection
-$app->database = function () use ($app): PDO {
+$app->database = function () use ($app): PdoInterface {
     if ('sqlite' == $app->config->get('database.driver')) {
         $dsn = 'sqlite:' . $app->config->get('database.dbname');
     } else {
