+++ src/Clear/Database/PdoExt.php
@@ -14,13 +14,25 @@ use Clear\Database\Event\{
 };
 use Psr\EventDispatcher\EventDispatcherInterface;
 use PDO;
+use InvalidArgumentException;
 
 /**
- * PdoExt extends PHP' internal PDO with PSR-14 Event Dispatcher
+ * PdoExt extends PHP's internal PDO with PSR-14 Event Dispatcher and Read/Write state
  */
-class PdoExt extends PDO implements PdoInterface
+final class PdoExt extends PDO implements PdoInterface
 {
+    const STATE_READ_ONLY   = 'r';
+    const STATE_READ_WRITE  = 'rw';
+    const STATE_UNAVAILABLE = '-';
+
     protected ?EventDispatcherInterface $dispatcher = null;
+
+    /**
+     * Database State - ReadWrite, ReadOnly and unavailable (NONE)
+     *
+     * @var string
+     */
+    private $state = self::STATE_READ_WRITE;
 
     public function __construct(string $dsn, string $username = '', string $passwd = '', array $options = [])
     {
@@ -40,7 +52,7 @@ class PdoExt extends PDO implements PdoInterface
     public function setEventDispatcher(?EventDispatcherInterface $dispatcher): void
     {
         $this->dispatcher = $dispatcher;
-        $this->setAttribute(PDO::ATTR_STATEMENT_CLASS, array('\Clear\Database\PdoStatementExt', array($this->dispatcher)));
+        $this->setAttribute(PDO::ATTR_STATEMENT_CLASS, array('\Clear\Database\PdoStatementExt', array($this, $this->dispatcher)));
     }
 
     /**
@@ -48,6 +60,10 @@ class PdoExt extends PDO implements PdoInterface
      */
     public function exec(string $statement): int|false
     {
+        if (!$this->canExecute($statement)) {
+            return false;
+        }
+
         $this->dispatch(new BeforeExec($statement));
         $res = parent::exec($statement);
         $this->dispatch(new AfterExec($statement, $res));
@@ -60,11 +76,52 @@ class PdoExt extends PDO implements PdoInterface
      */
     public function query(string $query, ?int $fetchMode = null, mixed ...$fetchModeArgs): PdoStatementExt|false
     {
+        if (!$this->canExecute($query)) {
+            return false;
+        }
         $this->dispatch(new BeforeQuery($query));
         $result = parent::query($query, $fetchMode, ...$fetchModeArgs);
         $this->dispatch(new AfterQuery($query, $result));
 
         return $result;
+    }
+
+    /**
+     * Returns the Database Read/Write state
+     *
+     * @return string
+     */
+    public function getState(): string
+    {
+        return $this->state;
+    }
+
+    /**
+     * Sets the Database Read/Write state
+     *
+     * @param string $state
+     * @return self
+     * @throws InvalidArgumentException on invalid state
+     */
+    public function setState(string $state): self
+    {
+        if (!in_array($state, [self::STATE_READ_WRITE, self::STATE_READ_ONLY, self::STATE_UNAVAILABLE], true)) {
+            throw new InvalidArgumentException('Invalid state provided');
+        }
+        $this->state = $state;
+
+        return $this;
+    }
+
+    public function canExecute(string $queryString)
+    {
+        if (self::STATE_READ_WRITE === $this->state) {
+            return true;
+        }
+        if (self::STATE_UNAVAILABLE === $this->state) {
+            return false;
+        }
+        return ! preg_match("/^(ALTER|CREATE|DELETE|DROP|GRANT|INSERT|RENAME|REVOKE|TRUNCATE|UPDATE)\s/i", trim($queryString));
     }
 
     private function dispatch(object $event): void
